Rename PepInscriptionContent to PepInscription and move to ui/
