+++ src/components/ui/PepInscription.spec.ts
@@ -1,13 +1,13 @@
 import { describe, it, expect } from 'vitest';
 import { mount } from '@vue/test-utils';
-import PepInscriptionContent from './PepInscriptionContent.vue';
+import PepInscription from './PepInscription.vue';
 import { inscriptionContentUrl } from '@/utils/inscriptions';
 
-describe('PepInscriptionContent', () => {
+describe('PepInscription', () => {
   const id = 'abc123i0';
 
   it('renders an <img> for image content types', () => {
-    const wrapper = mount(PepInscriptionContent, {
+    const wrapper = mount(PepInscription, {
       props: { id, contentType: 'image/png' }
     });
     const img = wrapper.find('img');
@@ -17,14 +17,14 @@ describe('PepInscriptionContent', () => {
   });
 
   it('renders a <video> for video content types', () => {
-    const wrapper = mount(PepInscriptionContent, {
+    const wrapper = mount(PepInscription, {
       props: { id, contentType: 'video/mp4' }
     });
     expect(wrapper.find('video').exists()).toBe(true);
   });
 
   it('renders an <iframe> with no allow-scripts when interactive=false (grid mode)', () => {
-    const wrapper = mount(PepInscriptionContent, {
+    const wrapper = mount(PepInscription, {
       props: { id, contentType: 'text/html', interactive: false }
     });
     const iframe = wrapper.find('iframe');
@@ -33,7 +33,7 @@ describe('PepInscriptionContent', () => {
   });
 
   it('enables allow-scripts on the sandboxed iframe only when interactive=true', () => {
-    const wrapper = mount(PepInscriptionContent, {
+    const wrapper = mount(PepInscription, {
       props: { id, contentType: 'text/html', interactive: true }
     });
     const iframe = wrapper.find('iframe');
@@ -41,14 +41,14 @@ describe('PepInscriptionContent', () => {
   });
 
   it('does not render an audio element in non-interactive (grid) mode', () => {
-    const wrapper = mount(PepInscriptionContent, {
+    const wrapper = mount(PepInscription, {
       props: { id, contentType: 'audio/mpeg', interactive: false }
     });
     expect(wrapper.find('audio').exists()).toBe(false);
   });
 
   it('falls back to a labeled placeholder for unknown content types', () => {
-    const wrapper = mount(PepInscriptionContent, {
+    const wrapper = mount(PepInscription, {
       props: { id, contentType: 'application/octet-stream' }
     });
     expect(wrapper.find('img').exists()).toBe(false);
